Use deprecated Moto decorators that patch socket
Boto3 v1.8+ replaced `requests` with `urllib3`, which broke the way Moto
was mocking components of Boto3.

The deprecated Moto decorators used in this change set mock at the
socket level, which allow it to bypass details about the HTTP client
used. This approach is not ideal, but more desirable than the
alternative of pinning Boto<1.8.