Move lint commands to own tox & Travis stage
This helps speed up the tox and Travis runs by running the lint commands
in parallel to the test commands. The lint commands need only be run
once for the project, not once for every support Python variation. This
reduces the necessary dependencies for running the tests without running
the lint commands. As linting doesn't require installing the library,
use tox option skip_install to save an additional small amount of time.