Configure setuptools using setup.cfg and declarative syntax (#56)
This moves the setuptools definition outside of setup.py to setup.cfg.
It replaces code with declarative non-code syntax. This reduces the
amount of custom code required to build and install the library.

Details on this setuptools feature is available at:

https://setuptools.readthedocs.io/en/latest/setuptools.html#configuring-setup-using-setup-cfg-files

> Setuptools allows using configuration files (usually setup.cfg) to
> define a package’s metadata and other options that are normally
> supplied to the setup() function (declarative config).
>
> This approach not only allows automation scenarios but also reduces
> boilerplate code in some cases.

Other popular projects including Django and setuptools itself use this
approach.